Adding viewDist argument Render3D. Adding getters and setters for Players in Game

+++ src/com/jamie/raycasting/entities/mobs/Mob.java
@@ -24,7 +24,7 @@ public abstract class Mob extends Entity
 
     // distances
     public double baseReach = 1;
-    public double viewDist = 4;
+    public int viewDist = 4;
 
     // actions
     public InputHandler input;
